bump libbpf-rs to 0.24.0+
This is quite a large commit as 0.24.0 contains several changes we need
to take into account:

- names are being changed to OsStr
- progs_iter() is being replaced by progs()
- prog("name") got removed in favor of the progs() iterator (and so we
need to manually look for the program name)
- commit c01e9b9 ("libbpf-cargo: Define maps as publicly accessible
members") upstream allows for a finer control of the rust lifetimes but
that means that we can not keep the tracing attach_prog as a static
global variable. It needs to be retrieved on-demand, but we ensure to
test it once to not retry several times when we know it will fail

Note that libbpf-rs 0.25.0 is also compatible with these changes, thus
marking the dependency as `>=`.

Fixes: #71

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>
Part-of: <https://gitlab.freedesktop.org/libevdev/udev-hid-bpf/-/merge_requests/205>