bpf: invert the have_struct_ops logic
Currently we detect if a bpf file contains a struct_ops program for
chosing whether we are going to use struct_ops or tracing. However, the
tracing loader is more restrictive than the struct_ops one, which means
that we currently can not have a generic program which just have a
probe.

Having a single probe and no events/report fixup is useful to query the
device or set it up into a special mode without actually changing the
code path after.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>
Part-of: <https://gitlab.freedesktop.org/libevdev/udev-hid-bpf/-/merge_requests/219>