bpf: read the report descriptor earlier in the process
We want to parse the report descriptor and inject it into the HID-BPF
object. However, this parsing and injecting needs to happen before we
call the probe. So we need to read the report descriptor file before
calling probe, and so this seems like the best place to do.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>
Part-of: <https://gitlab.freedesktop.org/libevdev/udev-hid-bpf/-/merge_requests/213>