rust: also make a lib out of our project
With the parsing of the report descriptor, we'll need a second binary
target for udev-hid-bpf for testing only. So we need to be able to call
functions from there into this new binary, so create the library now as
the code is small enough.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>
Part-of: <https://gitlab.freedesktop.org/libevdev/udev-hid-bpf/-/merge_requests/213>