test/btf: when generating the ctypes struct, ensure the size is correct
The kernel likes to make use of `__attribute__(packed)`, which means
that we sometimes don't have the correct alignment in C. This leads to a
wrong size of the generated struct.

When generating the struct, ensure we have the correct size by
generating a temporary struct and comparing the size with what's
provided by btf. We need to use a new temporary struct because the
`_fields_` attribute is final.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>
Part-of: <https://gitlab.freedesktop.org/libevdev/udev-hid-bpf/-/merge_requests/214>