test: add hid-rdesc-to-c-struct rust tool
When testing the `HID_REPORT_DESCRIPTOR` functionality, we need to
inject the report descriptor as the C representation. Instead of
manually re-implementing the functionality in python and keep 2
implementations which might differ in the long run, it's simpler to
create a new binary that will be called from python which just takes the
raw report descriptor and convertis it into the expected C-struct.

We use base64 encoding between the 2 tools.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>
Part-of: <https://gitlab.freedesktop.org/libevdev/udev-hid-bpf/-/merge_requests/213>