test: inject the given report descriptor in the bpf.o
We are now ready to inject the report descriptor in the bpf object by
calling hid-rdesc-to-c-struct to get the raw value to inject and do the
memmove to set it.

If the tool fails to parse the given report descriptor we reset it to 0
to have a clean state in each new test.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>
Part-of: <https://gitlab.freedesktop.org/libevdev/udev-hid-bpf/-/merge_requests/213>