bpf: add hid_usages.h
This is an auto-generated copy of the in-kernel linux/hid.h for the
usages part. We have to duplicate it because we can not include that
kernel header directly in the bpf objects.

It is generated by the udev-hid-bpf tool through a jinja template.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>
Part-of: <https://gitlab.freedesktop.org/libevdev/udev-hid-bpf/-/merge_requests/213>