bpf: check that the HID_REPORT_DESCRIPTOR type is correct
Before injecting the binary report descriptor, we can check that we are
injecting proper binary data in the bpf by looking at the BTF
information. This way, we can prevent injecting a wrong data in a BPF
object compiled with an incompatible struct.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>
Part-of: <https://gitlab.freedesktop.org/libevdev/udev-hid-bpf/-/merge_requests/213>